+++ src/ralph/backends/storage/fs.py
@@ -6,6 +6,7 @@ import sys
 from pathlib import Path
 
 from ralph.defaults import FS_STORAGE_DEFAULT_PATH
+from ralph.utils import now
 
 from ..mixins import HistoryMixin
 from .base import BaseStorage
@@ -59,11 +60,7 @@ class FSStorage(HistoryMixin, BaseStorage):
         logger.debug("Found %d archives", len(archives))
 
         if new:
-            archives = set(archives) - set(
-                entry.get("id")
-                for entry in filter(self.fetched_from_backend, self.history)
-            )
-
+            archives = set(archives) - self.get_command_history(self.name, "fetch")
             logger.debug("New archives: %d", len(archives))
 
         for archive in archives:
@@ -93,7 +90,7 @@ class FSStorage(HistoryMixin, BaseStorage):
                 "id": name,
                 "filename": details.get("filename"),
                 "size": details.get("size"),
-                "fetched_at": self.get_current_iso_time(),
+                "fetched_at": now(),
             }
         )
 
@@ -122,6 +119,6 @@ class FSStorage(HistoryMixin, BaseStorage):
                 "id": name,
                 "filename": details.get("filename"),
                 "size": details.get("size"),
-                "pushed_at": self.get_current_iso_time(),
+                "pushed_at": now(),
             }
         )